+++ api.py
@@ -6,18 +6,18 @@ import nest_asyncio
 from flask import Flask, jsonify, render_template, request
 
 import detection
-from detection import DetectionData, DetectionSettings
+from api_versions.v1 import v1
+from api_versions.v2 import v2
 from utils.logging import main_logger
 from utils.registry import DECISION_STRATEGIES, DETECTION_METHODS
 
-# __import__('IPython').embed()
 nest_asyncio.apply()
 
 # The storage interface for the sessions
 session_storage = detection.session_storage
 
 # Instantiate a logger for this HTTP API
-logger = main_logger.getChild('api')
+logger = main_logger.getChild("api")
 
 # Initiate Flask app
 app = Flask(__name__)
@@ -33,30 +33,9 @@ def shutdown_server():
     os._exit(0)
 
 
-# DEPRECATED
-@app.route("/api/v1/url", methods=["POST"])
-def check_url_old():
-    json = request.get_json()
 
-    res = detection.test_old(DetectionData.from_json(json))
-
-    return res.to_json_str_old()
-
-
-@app.route("/api/v2/url", methods=["POST"])
-def check_url():
-    json = request.get_json()
-    json_data = json["data"]
-    json_settings = json["settings"]
-
-    res = detection.test(
-        DetectionData.from_json(json_data), DetectionSettings.from_json(json_settings)
-    )
-
-    return res.to_json_str()
-
-
-@app.route("/api/v1/url/state", methods=["POST"])
+@v1.route("/url/state", methods=["POST"])
+@v2.route("/url/state", methods=["POST"])
 def get_url_state():
     json = request.get_json()
     url = json["URL"]
@@ -70,7 +49,8 @@ def get_url_state():
     return jsonify(result)
 
 
-@app.route("/api/v1/capabilities", methods=["GET"])
+@v1.route("/capabilities", methods=["GET"])
+@v2.route("/capabilities", methods=["GET"])
 def get_available_capabilities():
     result = [
         {
@@ -80,6 +60,10 @@ def get_available_capabilities():
     ]
     print(list(DECISION_STRATEGIES.items())[0])
     return jsonify(result)
+
+
+app.register_blueprint(v1, url_prefix="/api/v1")
+app.register_blueprint(v2, url_prefix="/api/v2")
 
 
 # Handle CTRL+C for shutdown
